requirements.txt
geographiclib==1.50
geopy==2.0.0
numpy==1.19.1
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
timezonefinder==4.4.1
XlsxWriter==1.3.3
./Eph/Whl/pyswisseph-2.5.1.post0-cp38-cp38-win_amd64.whl ; sys_platform == 'win32' and python_version == '3.8' and (platform_machine == "x86_64" or platform_machine == "AMD64")

./Eph/Whl/pyswisseph-2.5.1.post0-cp37-cp37m-win_amd64.whl ; sys_platform == 'win32' and python_version == '3.7' and (platform_machine == "x86_64" or platform_machine == "AMD64")

./Eph/Whl/pyswisseph-2.5.1.post0-cp36-cp36m-win_amd64.whl ; sys_platform == 'win32' and python_version == '3.6' and (platform_machine == "x86_64" or platform_machine == "AMD64")

./Eph/Whl/pyswisseph-2.5.1.post0-cp38-cp38-win32.whl ; sys_platform == 'win32' and python_version == '3.8' and (platform_machine == "x86_32" or platform_machine == "AMD32")

./Eph/Whl/pyswisseph-2.5.1.post0-cp37-cp37m-win32.whl ; sys_platform == 'win32' and python_version == '3.7' and (platform_machine == "x86_32" or platform_machine == "AMD32")

./Eph/Whl/pyswisseph-2.5.1.post0-cp36-cp36m-win32.whl ; sys_platform == 'win32' and python_version == '3.6' and (platform_machine == "x86_32" or platform_machine == "AMD32")
pyswisseph ; sys_platform != 'win32'
